% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clip_parXtreem}
\alias{clip_parXtreem}
\title{clip parXtreem table by date}
\usage{
clip_parXtreem(x, deploy = NA, recover = NA)
}
\arguments{
\item{x}{tibble, parXtreem}

\item{deploy}{POSIXt or NA, if not NA, clip data before this time}

\item{recover}{POSIXt or NA, if not NA, clip data before this time}
}
\value{
tibble
}
\description{
clip parXtreem table by date
}
