% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{draw_plot}
\alias{draw_plot}
\title{Plot PAR data as line plot}
\usage{
draw_plot(
  x = read_parXtreem(),
  main = "PAR",
  xlabel = "Date",
  ylabel = "PAR (umols / m squared / second)"
)
}
\arguments{
\item{x}{tibble of parXtreem data}

\item{main}{character, title}

\item{xlabel}{character, title of xaxis}

\item{ylabel}{character, title of yaxis}
}
\value{
ggplot2 object
}
\description{
Plot PAR data as line plot
}
