% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_parXtreem}
\alias{read_parXtreem}
\title{read parXtreem data file}
\usage{
read_parXtreem(filename = example_filename(), deploy = NA, recover = NA)
}
\arguments{
\item{filename}{character, the name of the file}

\item{deploy}{POSIXt or NA, if not NA, clip data before this time}

\item{recover}{POSIXt or NA, if not NA, clip data before this time}
}
\value{
tibble
}
\description{
read parXtreem data file
}
